More type hackery.
The upshot of this is that that which I am trying to do is probably too
difficult to accomplish using rusty techniques. It will be more
practical to emulate a more high-level programming system using Rc's and
RefCells. This is a little sad but the fact is that the essence of the
design that I am going for is that extreme runtime flexibility is
desired with lots of shared mutable state (activated objects talking to
each other by writing to memory that each other see) -- this is the only
design I am aware of that will work and it obviously can only be
expressed using Rc's and RefCells. Most lifetime annotations should go
away from this part of the code as the code moves in this direction.

Why is this the only kind of design that will work? For example, why
can't an overarching mutator call the functions of an rtor's children
with the data that they need, instead of the children passing references
between each other? The reason is that I want it to be possible to write
children that do weird, complex things to each other that the
runtime/interpreter would not understand about. This is flexible because
children are to be user-provided. Obviously more explanation is required
to fully justify this argument but such explanation is beyond the scope
of this commit message.