Continue rewriting the RtorInterface trait.
The latest change is to separate it into an immutable part and a mutable
part. There are several speculative ways of specifying level-like
information here and they are changing as the code is written. This code
is aiming for extreme generality and so is not likely to have a
particularly short path to an MVP (although it is still my intention to
take the shortest path possible to MVP, as usual).