Add BiFun; find design flaw.
The test result just promoted was semi-wrong. The decision now is that
immut_provide_unique really should be the same levels as immut_provide,
just uniquified and finitified thereby. There should always be a
self-consistent view of the levels. This is because I do not want
immut_provide_unique to have to do recursive analyses. I want the
interface information to be manifest in the iface declaration and in
that which is exposed by the iface declaration, not in the internals
(e.g., connections, non-exposed rtors) hidden behind the iface
declaration. However, it is also a hard requirement that we send "not
one extra message" (think C++, "not one extra instruction" in
IsWorseReallyBetter; then replace "instruction" with "RPC" or "message")
-- why? because I think of rtor API calls as messages -- why? because
I do _not_ want a system where you write the word
"federated" and "everything changes under the hood" (the phrase I
believe used by Marten? or Soroush or both to describe federated
execution) -- why? because I am interested in "simpler performance
models" (IsWorseReallyBetter) and that means _not_ abstracting over
profound performance characteristics (think asymptotics; asymptotic
differences are, perhaps surprisingly if you don't think about it, _not_ at
all hard to come by in the case of different implementations of an
entire system -- why? because the number of variables is huge and there
will always be one variable wrt which the asymptotics change; and also
there are many cost functions -- execution time, compilation time/code
size, space, ...) as you must do when you completely swap
implementations -- and why else? because "everything changes under the
hood" means double work -- two entirely seperate designs to maintain for the
under-the-hood -- and even worse than double work, making the two
separate designs talk to each other across ill-defined abstraction
boundaries -- and why else? because "everything
changes under the hood" means bizarre and embarassing corner cases --
think TopLevelArtifacts.lf.

And all of this is to say: Not one extra API call because an API call
must be convertible more or less one-to-one with a message over a
geographically distributed network. And not one extra API call (at
runtime, from one rtor to another) means more information is required in
the iface section of ir.