Augment the ifaces. See previous commit message.
It was clear in the previous commit that more information is required to
specify when it is and is not okay to increment the level. This is
because the local information exposed by the ifaces of the exposed rtors
is insufficient to specify when the level needs to increase, or
equivalently, when a notification is required in order for the srtor to
advance. In other words, when a notification is required is a property
of the entire srtor, that is, of the composition of the rtors making up
the srtor.

In recent days it further has become necessary that the markers used to
indicate incrementation of the level, which previously were merely
marked as "spacers" to evoke creation of a gap in the level, might be
given more meaning than merely that of child exposed rtors that accept
zero messages but that cause a gap in level. In this sense my decision
to mark them as separate variants of an enum rather than as sugar for
something else was probably a lucky guess at what would be the right
design. In their current state, these markers, now renamed as "Notify",
may be considered the _only_ cause for level incrementation, and should
correspond one-to-one with coordination messages/API calls; furthermore,
this opens up opportunity to optimize by allowing the upstream reactor
to consume these "Notify" markers provided via a connection to a
downstream reactor; by consuming the markers, it promises to notify when
ready. If it does not consume the markers, the parent reactor must
notify. The advantage of not consuming the markers is that direct
peer-to-peer communication can reduce network latency by half and avoid
the bandwidth bottleneck/single-point-of-failure issues that are
possible when going through the parent.

One way for a marker not to be consumed is for it simply not to be
involved in a connection. This happens when an iface consists
of some rtors and some Notifys, and the rtors are explicitly selected in
connections; Notifys, by contrast, cannot be explicitly selected because
there is no way to express that in the ir.

When two srtors are connected, the top-level Notifys of the left and
right rtors will not be handled by the children, so in order for the
induction step to work the two srtors must handle those Notifys.

Determination of responsibility for Notifys is, like computation of
concrete level assignment, the responsibility of the RtorComptimes, not
the RtorIfaces. Note that this is no problem because the previous
paragraph stipulates that the induction step only need involve the
top-level-notifys, which is manifest in the RtorIfaces.

---

The latest thought on consuming notifys is the following:
- To know whether it is safe to consume a notify, its nesting must be
  known. It is not safe to consume a notify that pertains to an
  rtor that is more shallowly-nested than the current one because such a
  notify could pertain to connections that precede the consuming rtor.
- If it is not safe to consume a notify, then it should be possible to
  defer the notify. The defer set should be part of the iterator.
- rtors that produce inputs should be able to view the defer set;
  generally if one is producing an input then the defer set should be
  empty. But this poses a design problem if the defer set is part of the
  iterator because in that case it would be part of the iterator over
  what the other rtor produced, not over what the current rtor produces.
  Probably the defer set should not be part of the iterator.