Implement Nesting.
This simple feature was alarmingly complex to add and makes it seem as
though I am doing something wrong. The complexity of adding more
lifetime annotations including in RtorIface is basically accidental; it
is obvious why it is necessary and not-obvious what design would be
better, but it is obvious that it could have been avoided by doing more
panic-prone things at runtime or by introducing yet more intermediate
structs and types. The awkward dance in moving between backing iterators
in ChainClone is largely about keeping the borrow checker happy while
breaking invariants only temporarily and so is basically just accidental
complexity. But I really wanted to use the move semantics to reuse
memory and ensure that transitions between the backing iterators of the
ChainClone were being made properly. The whole thing feels like a little
code golf exercise and the outcome is not pleasing.