Solve?? own-vs-borrow problem with backing iters.
It feels a little silly to be relieved to have solved such a trivial
problem which would not have existed if this were implemented in a more
"dynamic" language than Rust. But this is what happens when you have to
decide what to reference count and what to own and not to own etc.
rather than having everything be memory managed in a uniform way
automatically. It also should be admitted that there is some true
essential complexity that cannot be blamed on the implementation
language of this API: namely, it is necessary to be very careful about
who modifies the iterators and who doesn't, and it is necessary to make
changes on the right copy in certain cases instead of on a clone. This
error-prone situation comes from my decision to make everything
cloneable despite crucial reliance on passing around mutable objects
(iterators); but the cloneability is hard to eliminate because it really
is desirable for the subparts of the ifaces to hold on to copies of the
information that they are granted.